Crud operations to admin controller
