+++ hiilipolkuapp.Server/Controllers/AuthorizationController.cs
@@ -1,0 +1,16 @@
+﻿using hiilipolkuapp.Server.Classes;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+
+namespace hiilipolkuapp.Server.Controllers
+{
+    [ApiController]
+    [Route("api/[controller]")]
+    public class AuthorizationController : ControllerBase
+    {
+        //[HttpGet]
+        //public async Task<ActionResult<User>> GetUser()
+        //{
+        //}
+    }
+}
